requirements.txt
certifi==2023.5.7
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==2.1
django-cors-headers==2.4.0
django-rest-swagger==2.2.0
django-seed==0.1.9
djangorestframework==3.8.2
Faker==0.9.0
filelock==3.12.0
idna==2.7
importlib-metadata==6.6.0
itypes==1.1.0
Jinja2==2.10
lxml==4.9.2
MarkupSafe==1.1.1
mysqlclient==1.3.13
openapi-codec==1.3.2
Pillow==9.5.0
pipenv==2023.5.19
platformdirs==3.5.1
python-dateutil==2.7.3
python-docx==0.8.7
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
typing_extensions==4.6.0
uritemplate==3.0.0
urllib3==1.23
virtualenv==20.23.0
virtualenv-clone==0.5.7
zipp==3.15.0
